Don't track rising/falling separately on Day2
Cute hack: use -1,1 to indicate direction, then multiply by previous
direction before replacing it.

The first time its zero and -1*0 == 1*0 == 0.
If the signs match, 1*1 == -1*-1 == 1.
If they mismatch, -1*1 == 1*-1 == -1.

So if we ever see -1, it's unsafe.

Also remove the accumulate, I decided it's ugly and this way we can
early-out the moment we find any unsafety.