Day5: ⭐️🌟
Works, needs clean.

Ended up adding a new structure to track the rules. Either its useful
and we could update part1 to use it instead of scanning a vector of
pairs. Or its overkill and we're tracking more than we need, ie we might
not need to keep both list of what is bigger AND smaller than each.

It looks like the input is almost guranteed to have rules for the pairs
we need to consider. I was worried we'd have to figure out transitive
orderings, but instead there's a single rule to answer any question in
the update list, just gotta find it.