Day9: ⭐️🌟
I had grander plans, but it works.
Could keep a feeList of empty blocks, but then I'd have to split it up
on partial allocations and re-insert according to remaining size, if
any. But this is fast enough, so ok.

Also lots of gross mixing of iterator and index math.