Cleanup
Now we have a common main() funciton, which expects a process() that takes
a stringstream. We get that by linking against a given day's solution.

The harness now reads the expected file and compares results, the intent
is to eventually have something better than redirecting to a file and
running diff.