Day11: ⭐️🌟
So while the literal interpretation of the question (modelling rows of
stones as actual vector of numbers) may work for 25 iterations, it
absolutely does not for 75. I got killed at something like 180GB or RAM
usage.

Noting that no number depends on any other, we can just see how many
stones the first one expands to for a given generation, and add that to
how many the next does. This way we don't keep all 64-200 trillion
numbers at once. Turns out this is still super slow, even though memory
use is under control.

So we just slap a memoization layer in front and it suddenly becomes
super tractable: 50ms to do part1 and part1 of example and my real
input!