Day15: ⭐️🌟
Part1 was done naively, exactly as the worked examples say one might do
on paper. The grid is the underlying datastructure, and we can push
a box if it can push the single box next to it in the direction of
pushing. Pushing into '.' always succeeds, and pushing into '#' always
fails. If the last box in a row/col can be moved, it swaps itself with
the space, opening a space for the previous box. And so on, until
the robot itself swaps itself with the new space right next to it.

That didnt work so well for Part2 - east/west pushes worked more or less
the same, but pushing vertically on [ or ] got annoying quickly.
The new approach is to read in positions of blockers and boxes, where
each box knows the coordinates of its left and right sides. We can now
look up a box by coord, whether thats the left or right side and see if
the box as a whole can move.

Horizontally works easy enough: when pushing from the left we have to
ensure the righthand side has space, and vice-versa.
Vertically we just see if the both sizes can fail: Trivially yes if both
are empty, or no if either is #.

If either or both sides push on another box we see if both are in turn
recursively moveable. However, once we have a 'yes', we cant immediately
do the move in case the other fork fails, resulting in zero movement.
Instead we append to a list of swaps, and only if the whole stack is
indeed moveable, do we apply the list of move commands and finally move
the robot.

Rather than porting part1 to the new method, just leave as originally
submitted since it works for that case.