Revisit day14.
I wasnt super satisfied with the "find the triangle" approach and wanted
to see if the floodfill would have eventually found something.

Original plan of waiting for "all" robots to be adjacent wouldn't work,
the clue even says "most" form a picture of a tree.

Looking for >=50% adjacent also wouldn't help, since the tree and the
box around it are disconnected: together they form "most" but neither
does individually.

So lets just look for the biggest adjoining blob of robots in max iters
(which seems to be width*height?) and assume that is the tree. IN my
case it is, 230/500 robots making up "tree" portion.