Day16: ⭐️🌟
Got the second star after a revisit.

First attempt was done as a depthsearch, looking for tendrils of the shortest length which
reach the goal. This was... slow, and I more or less have up afte an
hour of runtime.

Some days later, I ended up adding path tracking to my Dijkstra. Now for
each current node in the exploration, we know the set of all positions
on the path to get there.

With that it was "just" a matter of allowing the walk to find all shortest nodes instead of
exiting on the first one - we track the "maybe shortest" and set it on
the first find. The next few pops on the prio queue can be expected
to have equal cost, and may even reach the goal - if so we accumulate
their seen paths into the overall vantage set.

Finally, when adding to the prio queue, rather than ignoring pushes of
greater or equal cost, we only drop _more_ expensive candidates.
Equivalent nodes of equal cost will have different histories to have
gotten to that point. So we merge the histories as well.

With that, we got our 2nd star and the runtime is like 2 seconds.