Day17: ⭐️🌟
Find the quine.

First we effectively disassemble the program to understand what it's
doing - we can see it's processing the lowest 3-bits at a time to
produce the output, and also mix into the upper bits.

So we find which 3 bits give us one desired output (there may be
several, so dont be greedy!) and see if appending some next-3-bits gives
us the next digit.

Recurse this until we've seen the whole program reflected back.