Day24  ⭐️🌟
Bit of a rewrite for part2. Rather than relaxing each wire until
everything settles, we try recover the schematics. Start by computing
a wire by recursing over its inputs until we terminate. This is another
way to solve P1.

With this, we can 'annotate' the schematic, by essentially taking the
definition of half (for bit0) and full (for bits 1-44) adders and
checking each net. For example, in a standard adder, the output bit 0 is
the XOR of each input's bit0. So we can search for which node does the
XOR of y00 and y00, and know the output is sum00. Similarly the AND of
those two bits is carry00, which is consumed by the full adder of bit
01.

Following this, we can find and verify all steps of the 45-bit adder
chain. There's facility to rename nets to something useful like "halfum"
or "carry" and dump the renames in a similar format to the input file.

This way we could run until some check failed, then manually eyeball and
swap the affected output nets. This happened exactly four times, and
this was our answer.

The current program doesnt determine the 4x net pairs to swap, but we
can at least re-check the ones derived manually.. This is less
satisfying, but ok for now.