Day20  ⭐️🌟
Same approach as before: find the shortest path, then walk it again
looking through walls to see if we land elsewhere ahead on the track in
less time than it would take to get there legitimately.

In p1, we used one step of the cadrinal compass directions for this
look-around.

Now, we explicitly build the "constellation", a kind of kernel to
convolve over the map as we walk it. Retrofitting part1 to use
a constellation of max manhattan distance 2 yields the same result
again, and using one with max distance 20 solves for part2.