Day21  ⭐️🌟
As suspected, took a bit of a rethink.

Instead of keeping all shortest numpad strings, and using those to
calculate all shortest of the next level, and using those to calculate
same for the next level... we can reason about individual transitions.

So we ask the numpad how many presses it takes to get from key A to key
B. It can calculate all shortest paths for each prev,next pair of keys
in those ask the next level, who does the same. Eventually the last
level (whether 2 or 25) simply returns the smallest number of presses.

The key is, each keypad can remember the best score when it asks the
next pad to move if from any source to any destination. Say keypad 17
needs to move from < to A.. The first time it will have to ask keypad16
for it's counts for each of the 3 combinations and take the smallest
- but the next time it can simply reuse the result and avoid much
recomputing.

This made part1 MUCH faster, and makes part2 tractable.